+++ baiducloud-python-sdk-cfw/baiducloud_python_sdk_cfw/models/cfw_rule.py
@@ -20,10 +20,13 @@ class CfwRule(AbstractModel):
         direction=None,
         source_address=None,
         dest_address=None,
+        domain_detect_mode=None,
+        applications=None,
         source_port=None,
         dest_port=None,
         action=None,
         description=None,
+        created_time=None,
     ):
         """
         Initialize CfwRule instance.
@@ -52,6 +55,12 @@ class CfwRule(AbstractModel):
         :param dest_address: 目的IP，网段或IP，例如：180.76.1.0/24 或 180.76.1.1，0.0.0.0/0表示所有
         :type dest_address: str (optional)
 
+        :param domain_detect_mode: 域名解析模式，当destAddress为域名时有效，取值 \\[ dns \\| fqdn ]
+        :type domain_detect_mode: str (optional)
+
+        :param applications: 应用，支持的应用协议有HTTP、SSL、FTP、 RDP、 SMTP、SSH、VNC
+        :type applications: List[str] (optional)
+
         :param source_port: 源端口，0-65535之间的整数，或区间，例如：8080，8080-8090，0-65535表示所有
         :type source_port: str (optional)
 
@@ -63,6 +72,9 @@ class CfwRule(AbstractModel):
 
         :param description: CFW规则的描述
         :type description: str (optional)
+
+        :param created_time: CFW规则的创建时间，标准UTC时间格式
+        :type created_time: str (optional)
         """
         super().__init__()
         self.cfw_id = cfw_id
@@ -73,10 +85,13 @@ class CfwRule(AbstractModel):
         self.direction = direction
         self.source_address = source_address
         self.dest_address = dest_address
+        self.domain_detect_mode = domain_detect_mode
+        self.applications = applications
         self.source_port = source_port
         self.dest_port = dest_port
         self.action = action
         self.description = description
+        self.created_time = created_time
 
     def to_dict(self):
         """
@@ -107,6 +122,10 @@ class CfwRule(AbstractModel):
             result['sourceAddress'] = self.source_address
         if self.dest_address is not None:
             result['destAddress'] = self.dest_address
+        if self.domain_detect_mode is not None:
+            result['domainDetectMode'] = self.domain_detect_mode
+        if self.applications is not None:
+            result['applications'] = self.applications
         if self.source_port is not None:
             result['sourcePort'] = self.source_port
         if self.dest_port is not None:
@@ -115,6 +134,8 @@ class CfwRule(AbstractModel):
             result['action'] = self.action
         if self.description is not None:
             result['description'] = self.description
+        if self.created_time is not None:
+            result['createdTime'] = self.created_time
         return result
 
     def from_dict(self, m):
@@ -149,6 +170,10 @@ class CfwRule(AbstractModel):
             self.source_address = m.get('sourceAddress')
         if m.get('destAddress') is not None:
             self.dest_address = m.get('destAddress')
+        if m.get('domainDetectMode') is not None:
+            self.domain_detect_mode = m.get('domainDetectMode')
+        if m.get('applications') is not None:
+            self.applications = m.get('applications')
         if m.get('sourcePort') is not None:
             self.source_port = m.get('sourcePort')
         if m.get('destPort') is not None:
@@ -157,4 +182,6 @@ class CfwRule(AbstractModel):
             self.action = m.get('action')
         if m.get('description') is not None:
             self.description = m.get('description')
+        if m.get('createdTime') is not None:
+            self.created_time = m.get('createdTime')
         return self