+++ baiducloud-python-sdk-cfw/test/cfw_client_test.py
@@ -16,7 +16,10 @@ class CfwClientTest(unittest.TestCase):
         HOST = b''
         AK = b''
         SK = b''
+
+        # ==== AK/SK 鉴权 ====
         config = BceClientConfiguration(credentials=BceCredentials(AK, SK), endpoint=HOST)
+
         self.client = CfwClient(config)
 
     def tearDown(self):